pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cwsf"
version = "0.1.0"
description = "Configurable Web Scraping Framework"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click",
    "httpx",
    "parsel",
    "pyyaml",
    "jsonschema",
    "watchdog",
    "apscheduler",
    "playwright",
]

[project.scripts]
cwsf = "cwsf.cli:entry_point"

[tool.setuptools.packages.find]
where = ["."]
include = ["cwsf*"]
